abstracted some allow list functions and handle dup primary key
